requirements.txt
asgiref==3.2.10
Django==3.1.2
djangorestframework==3.12.1
pytz==2020.1
sqlparse==0.3.1
psycopg2==2.7.5
psycopg2-binary==2.7.5
Faker==8.1.2
requests==2.22.0
